feat[animation]: Animation controller working
* Optimized the code for componentTransform: the recursive call in CalculateMatrices isn't needed (and it destroys performane inmensavely) if the UpdateTrasnformMatrices function already calls the whole tree recursively from top down order for the object to update.

+++ Source/DataModels/Animation/AnimationController.h
@@ -24,7 +24,7 @@ private:
 	float3 Interpolate(const float3& first, const float3& second, float lambda);
 	Quat Interpolate(const Quat& first, const Quat& second, float lambda);
 
-	unsigned int currentTime;
+	float currentTime;
 	bool loop;
 	bool play;
 	std::shared_ptr<ResourceAnimation> resource;
